refactor(dot11): relocate test module into tests directory
